feat: Pass useSelectedCommittee hook to separate file
