+++ app/src/App.js
@@ -1,6 +1,6 @@
-import React, { useState, useMemo, useCallback } from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
-import { useAragonApi, usePath, useAppState } from '@aragon/api-react'
+import { useAragonApi } from '@aragon/api-react'
 import {
   Main,
   Button,
@@ -22,48 +22,7 @@ import NoCommittees from './screens/NoCommittees'
 import Committees from './screens/Committees'
 import CommitteeDetails from './screens/CommitteeDetails'
 
-const COMMITTEE_ID_PATH_RE = /^\/committee\/(0x[a-fA-F0-9]{40})\/?$/
-const NO_COMMITTEE_ADDRESS = null
-
-const idFromPath = path => {
-  if (!path) {
-    return NO_COMMITTEE_ADDRESS
-  }
-  const matches = path.match(COMMITTEE_ID_PATH_RE)
-  return matches ? matches[1] : NO_COMMITTEE_ADDRESS
-}
-
-export const useSelectedCommittee = committees => {
-  const [path, requestPath] = usePath()
-  const { appState } = useAragonApi()
-
-  const { isSyncing } = appState
-  // The memoized proposal currently selected.
-  const selectedCommittee = useMemo(() => {
-    const id = idFromPath(path)
-
-    // The `isSyncing` check prevents a proposal to be
-    // selected until the app state is fully ready.
-    if (isSyncing || id === NO_COMMITTEE_ADDRESS) {
-      return null
-    }
-
-    return committees.find(committee => committee.address === id) || null
-  }, [path, isSyncing, committees])
-
-  const selectCommittee = useCallback(
-    committee => {
-      requestPath(
-        committee === NO_COMMITTEE_ADDRESS
-          ? ''
-          : `/committee/${committee.address}/`
-      )
-    },
-    [requestPath]
-  )
-
-  return [selectedCommittee, selectCommittee]
-}
+import useSelectedCommittee from './hooks/useSelectedCommitee'
 
 const App = () => {
   const theme = useTheme()
